requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
google-genai==0.3.0
firebase-admin==6.2.0
geopy==2.4.0
requests==2.31.0
websockets==12.0
python-socketio==5.10.0
pytest==7.4.3
pytest-asyncio==0.21.1
